+++ tests/functional/package_test/test_regression_getall_wrapper.py
@@ -16,10 +16,10 @@ from codeCheckerDBAccess.ttypes import ReportFilter
 from codeCheckerDBAccess.ttypes import SortMode
 from codeCheckerDBAccess.ttypes import SortType
 
-from test_utils.debug_printer import print_run_results
-from test_utils.thrift_client_to_db import CCViewerHelper
-from test_utils.thrift_client_to_db import get_all_run_results
-from test_utils.result_compare import find_all
+from libtest.debug_printer import print_run_results
+from libtest.thrift_client_to_db import CCViewerHelper
+from libtest.thrift_client_to_db import get_all_run_results
+from libtest.result_compare import find_all
 
 
 class RunResults(unittest.TestCase):
