+++ tests/functional/package_test/test_suppress.py
@@ -11,9 +11,9 @@ import unittest
 
 from codeCheckerDBAccess.ttypes import ReportFilter
 
-from test_utils.debug_printer import print_run_results
-from test_utils.thrift_client_to_db import CCViewerHelper
-from test_utils.thrift_client_to_db import get_all_run_results
+from libtest.debug_printer import print_run_results
+from libtest.thrift_client_to_db import CCViewerHelper
+from libtest.thrift_client_to_db import get_all_run_results
 
 
 class Suppress(unittest.TestCase):
