+++ tests/functional/package_test/test_suppress_file_cmdline.py
@@ -9,7 +9,7 @@ import json
 import os
 import unittest
 
-from test_utils.thrift_client_to_db import CCViewerHelper
+from libtest.thrift_client_to_db import CCViewerHelper
 
 
 class RunResults(unittest.TestCase):
