+++ tests/libtest/debug_printer.py
