+++ tests/libtest/result_compare.py
